Testes

+++ clinica-medica-comum/src/main/java/br/edu/imepac/comum/dtos/funcionario/FuncionarioRequest.java
@@ -14,7 +14,7 @@ public class FuncionarioRequest {
     private String usuario;
     private String senha;
     private String nome;
-    private Integer idade;
+    private Long idade;
     private char sexo;
     private String cpf;
     private String rua;
