+++ clinica-medica-comum/src/main/java/br/edu/imepac/comum/repositories/ConsultaRepository.java
@@ -4,4 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ConsultaRepository extends JpaRepository<Consulta, Integer> {}
+public interface ConsultaRepository extends JpaRepository<Consulta, Long>{
+
+}
+
