+++ clinica-medica-comum/src/main/java/br/edu/imepac/comum/repositories/ConvenioRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ConvenioRepository extends JpaRepository<Convenio, Integer> {}
+public interface ConvenioRepository extends JpaRepository<Convenio, Long> {}
