+++ clinica-medica-comum/src/main/java/br/edu/imepac/comum/repositories/PacienteRepository.java
@@ -2,4 +2,4 @@ package br.edu.imepac.comum.repositories;
 import br.edu.imepac.comum.models.Paciente;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PacienteRepository extends JpaRepository<Paciente, Integer> {}
+public interface PacienteRepository extends JpaRepository<Paciente, Long> {}
